Create visualise_tarql.py
This routine extracts the CONSTRUCT{} part of a TARQL file and outputs as a turtle file.  It is used in conjunction with the turtle editor viewer ( https://www.semantechs.co.uk/turtle-editor-viewer/ ) to visualise a tarql file.  Several TARQL files can be aggregated to evaluate the whole model.  Evaluation can include developing and testing SPARQL queries, and also checking inferences using the "Show Facts" button in the turtle viewer app.